+++ Malloc.h
@@ -16,7 +16,7 @@ The files Malloc.[ch] provide several functionalities:
 The module defines several sets of routines:
 
 1.  void *Malloc(size_t s)
-    void *Calloc(int n, size_t s)
+    void *Calloc(size_t n, size_t s)
     void *Realloc(void *p, size_t s)
     void Free(void *p)
 
@@ -72,19 +72,19 @@ The module defines several sets of routines:
   the time.
 *****/
 
-#define	Malloc(s)	(_leak_malloc(1, (size_t)(s), __FILE__, __LINE__))
-#define	Calloc(n,s)	(_leak_calloc(1, (n), (size_t)(s), __FILE__, __LINE__))
-#define	Realloc(p,s)	(_leak_realloc(1, (void *)(p), (size_t)(s), __FILE__, __LINE__))
-#define	TryMalloc(s)	(_leak_malloc(0, (size_t)(s), __FILE__, __LINE__))
-#define	TryCalloc(n,s)	(_leak_calloc(0, (n), (size_t)(s), __FILE__, __LINE__))
-#define	TryRealloc(p,s)	(_leak_realloc(0, (void *)(p), (size_t)(s), __FILE__, __LINE__))
+#define	Malloc(s)	(_leak_malloc(1, (s), __FILE__, __LINE__))
+#define	Calloc(n,s)	(_leak_calloc(1, (n), (s), __FILE__, __LINE__))
+#define	Realloc(p,s)	(_leak_realloc(1, (void *)(p), (s), __FILE__, __LINE__))
+#define	TryMalloc(s)	(_leak_malloc(0, (s), __FILE__, __LINE__))
+#define	TryCalloc(n,s)	(_leak_calloc(0, (n), (s), __FILE__, __LINE__))
+#define	TryRealloc(p,s)	(_leak_realloc(0, (void *)(p), (s), __FILE__, __LINE__))
 #define	Free(p)		(_leak_free((void *)(p), __FILE__, __LINE__))
 
 #define	new(type)	((type *)Malloc(sizeof (type)))
 #define	new_string(s)	(_new_string((s), __FILE__, __LINE__))
 
 extern void *_leak_malloc(int chk, size_t size, const char *fname, int l_nmb);
-extern void *_leak_calloc(int chk, int n, size_t size, const char *fname, int l_nmb);
+extern void *_leak_calloc(int chk, size_t n, size_t size, const char *fname, int l_nmb);
 extern void *_leak_realloc(int chk, void *addr, size_t size, const char *fname, int l_nmb);
 extern void _leak_free(void *addr, const char *fname, int l_nmb);
