+++ text.h
@@ -15,19 +15,19 @@ struct text {
 				   part of a chunk; sorted and updated by
 				   Pass 2
 				*/
-	unsigned int tx_start;	/* positions in Token_Array[] for the text */
-	unsigned int tx_limit;
-	unsigned int tx_nl_start;/* possibly newline pointer for pass2 */
-	unsigned int tx_nl_limit;
+	size_t tx_start;	/* positions in Token_Array[] for the text */
+	size_t tx_limit;
+	size_t tx_nl_start;	/* possibly newline pointer for pass2 */
+	size_t tx_nl_limit;
 };
 
 struct position {
 	/* position of first and last token of a chunk */
 	struct position *ps_next;
 	int ps_type;		/* first = 0, last = 1 */
-	unsigned int ps_tk_cnt;	/* in tokens; set by add_run()
+	size_t ps_tk_cnt;	/* in tokens; set by add_run()
 				   in Read_Input_Files() */
-	unsigned int ps_nl_cnt;	/* same, in line numbers;set by Retrieve_Runs(),
+	size_t ps_nl_cnt;	/* same, in line numbers;set by Retrieve_Runs(),
 				   used by Show_Runs(), to report line numbers
 				*/
 };
@@ -43,5 +43,5 @@ extern int Next_Text_Token_Obtained(enum Pass pass);
 extern void Close_Text(enum Pass pass, struct text *txt);
 
 #ifdef	DB_NL_BUFF
-extern void db_print_nl_buff(unsigned int start, unsigned int limit);
+extern void db_print_nl_buff(size_t start, size_t limit);
 #endif
